requirements.txt
python-dotenv
requests
pytube
yt-dlp